+++ NEWS.md
@@ -1,6 +1,6 @@
 # implyr 0.2.4.999
 
-* Added more SQL translations including lubridate functions
+* Added more SQL translations including stringr and lubridate functions
 * Updated for compatibility with dplyr 0.8.3
 * Fixed errors when `na.rm = TRUE` is specified in some aggregate functions
 * Made minor bugfixes and improvements